requirements.txt
exceptiongroup 1.2.2
grpcio         1.68.0
grpcio-tools   1.68.0
iniconfig      2.0.0
packaging      24.2
pip            24.3.1
pluggy         1.5.0
protobuf       5.28.3
pytest         8.3.4
pytest-grpc    0.8.0
setuptools     59.6.0
tomli          2.2.1

